backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pika==1.3.2
geopy==2.4.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
celery==5.3.4
requests==2.31.0

